Further Code Review fixes: strong signing, moved MemoryFileCache, documentation
